UserProfileView Updated
